Refactor `Tone` to use `Double` for raw values
- Change the underlying type of the `Tone` value class from `Int` to `Double` to improve precision.
- Update `Tone` arithmetic operators, mapping functions, and range constants to support `Double`.
- Refactor `FContextExt.kt` to remove redundant `toDouble()` and `toInt()` conversions when calculating colors and contrast.
- Update `backgroundToneCalculators` to use `Double` literals for brightness and step calculations.

+++ app/model/src/commonMain/kotlin/org/hnau/commons/app/model/theme/color/Tone.kt
@@ -5,7 +5,7 @@ import kotlin.jvm.JvmInline
 
 @JvmInline
 value class Tone private constructor(
-    val raw: Int,
+    val raw: Double,
 ) : Comparable<Tone> {
 
     override fun compareTo(
@@ -15,14 +15,14 @@ value class Tone private constructor(
     )
 
     inline fun map(
-        transform: (Int) -> Int,
+        transform: (Double) -> Double,
     ): Tone = create(
         raw = transform(raw),
     )
 
     inline fun combineWith(
         other: Tone,
-        combine: (Int, Int) -> Int,
+        combine: (Double, Double) -> Double,
     ): Tone = map { tone ->
         combine(tone, other.raw)
     }
@@ -31,32 +31,32 @@ value class Tone private constructor(
         other: Tone,
     ): Tone = combineWith(
         other = other,
-        combine = Int::plus,
+        combine = Double::plus,
     )
 
     operator fun minus(
         other: Tone,
     ): Tone = combineWith(
         other = other,
-        combine = Int::minus,
+        combine = Double::minus,
     )
 
     operator fun times(
         factor: Number,
     ): Tone = map { tone ->
-        tone * factor.toInt()
+        tone * factor.toDouble()
     }
 
     operator fun div(
         factor: Number,
     ): Tone = map { tone ->
-        tone / factor.toInt()
+        tone / factor.toDouble()
     }
 
     companion object {
 
         fun create(
-            raw: Int
+            raw: Double
         ): Tone {
             requireInRange(
                 valueLabel = "Tone",
@@ -78,10 +78,10 @@ value class Tone private constructor(
         val max: Tone
             get() = Tone(maxRaw)
 
-        private val minRaw: Int
-            get() = 0
+        private val minRaw: Double
+            get() = 0.0
 
-        private val maxRaw: Int
-            get() = 100
+        private val maxRaw: Double
+            get() = 100.0
     }
 }
